backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
openai
pymysql
websockets
funasr
modelscope
torch
torchaudio
librosa
soundfile
numpy
edge-tts
python-jose[cryptography]
langchain-openai
langchain-core
